data/preparation/cognitive/requirements.txt
azure-cognitiveservices-vision-customvision~=1.0.0
python-dotenv~=0.13.0